requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
aiofiles==23.2.1

# OCR Model dependencies
transformers>=4.46.0
torch==2.1.2
torchvision==0.16.2
accelerate>=0.26.1
pillow>=10.2.0
numpy<2.0
qwen-vl-utils>=0.0.5

# PDF processing
PyMuPDF==1.23.21
pdf2image==1.17.0

# Image processing
opencv-python-headless==4.9.0.80

# API utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
typing-extensions==4.9.0

# Background tasks and scheduling
apscheduler==3.10.4

# Monitoring and logging
structlog==24.1.0
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Optional: Cloud storage (for future GCS integration)
# google-cloud-storage==2.14.0
